refactor : pr 리뷰 반영 #2
