+++ src/main/java/com/sendback/domain/user/repository/UserRepository.java
@@ -1,13 +1,13 @@
-package com.sendback.domain.users.repository;
+package com.sendback.domain.user.repository;
 
-import com.sendback.domain.users.entity.User;
+import com.sendback.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UsersRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     User findByNickname(String nickname);
     Optional<User> findBySocialId(String socialId);
 }
