+++ src/main/java/com/sendback/global/exception/BaseException.java
@@ -8,7 +8,7 @@ public class BaseException extends RuntimeException{
     private final ExceptionType exceptionType;
 
     public BaseException(final ExceptionType exceptionType) {
-        super(exceptionType.message());
+        super(exceptionType.getMessage());
         this.exceptionType = exceptionType;
     }
